refactor: swap urls and pathbuilders location
urls is needed for client but pathBuilders is only used for development
